refactor: move Resources from apt directory
